add verification messages to the post delete button and tournament delete buttons

+++ frontend/src/components/forumComponents/posts/PostDetail.js
@@ -17,7 +17,9 @@ class PostDetail extends Component {
   }
 
   onDelete() {
-    this.props.deletePost(this.props.post._id);
+    if (window.confirm("Are you sure you would like to delete this post?")) {
+      this.props.deletePost(this.props.post._id);
+    }
   }
 
   renderDeleteButton() {
